+++ src/index.ts
@@ -3,3 +3,5 @@ export * from "./parsers/csv-parser.ts";
 export * from "./parsers/json-parser.ts";
 
 export * from "./adapters/papaparse.adapter.ts";
+
+export * from "./converters/index.ts";